# nodejs-server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Estuary API
  description: This is the API for the Estuary application.
  termsOfService: http://estuary.tech
  contact:
    name: API Support
    url: https://docs.estuary.tech/feedback
  license:
    name: Apache 2.0 Apache-2.0 OR MIT
    url: https://github.com/application-research/estuary/blob/master/LICENSE.md
  version: 0.0.0
servers:
- url: //api.estuary.tech/
security:
- bearerAuth: []
paths:
...
  /public/deals/failures:
    get:
      tags:
      - deals
      summary: Get storage failures
      description: This endpoint returns a list of storage failures
      operationId: publicDealsFailuresGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
  /public/info:
    get:
      tags:
      - public
      summary: Get public node info
      description: This endpoint returns information about the node
      operationId: publicInfoGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
...
  /public/miners:
    get:
      tags:
      - public
      - net
      summary: Get all miners
      description: This endpoint returns all miners
      operationId: publicMinersGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
  /public/miners/deals/{miner}:
    get:
      tags:
      - public
      - miner
      summary: Get all miners deals
      description: This endpoint returns all miners deals
      operationId: publicMinersDealsMinerGET
      parameters:
      - name: miner
        in: path
        description: Filter by miner
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: ignore-failed
        in: query
        description: Ignore Failed
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
  /public/miners/failures/{miner}:
    get:
      tags:
      - public
      - net
      summary: Get all miners
      description: This endpoint returns all miners
      operationId: publicMinersFailuresMinerGET
      parameters:
      - name: miner
        in: path
        description: Filter by miner
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
  /public/miners/stats/{miner}:
    get:
      tags:
      - public
      - miner
      summary: Get miner stats
      description: This endpoint returns miner stats
      operationId: publicMinersStatsMinerGET
      parameters:
      - name: miner
        in: path
        description: Filter by miner
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
  /public/miners/storage/query/{miner}:
    get:
      tags:
      - deals
      summary: Query Ask
      description: This endpoint returns the ask for a given CID
      operationId: publicMinersStorageQueryMinerGET
      parameters:
      - name: miner
        in: path
        description: CID
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Deals
...
  /public/net/peers:
    get:
      tags:
      - public
      - net
      summary: Net Peers
      description: This endpoint is used to get net peers
      operationId: publicNetPeersGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
  /public/stats:
    get:
      tags:
      - public
      summary: Public stats
      description: This endpoint is used to get public stats.
      operationId: publicStatsGET
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/util.HttpError'
      x-swagger-router-controller: Public
...
components:
  schemas:
...
    util.HttpError:
      type: object
      properties:
        code:
          type: integer
        details:
          type: string
        reason:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header
...